Use table format in torrent-get by default if supported
Transmission 3.0 (RPC version 16) introduced a "table format" when creating a
reply to torrent-get, which uses arrays in order to avoid duplicated key names
in JSON objects and thus leads to smaller JSON payloads.

Note that we're still recreating the old "objects" representation in memory, as
all users of RequestInfo() use the returned TJSONObject directly and expect it
to be in the old format.

In my Transmission instance with around 2000 torrents, this reduces the reply
size from 2560 KiB to 1895 KiB. Maybe not a big win, but still.