Implement gzip-compressed HTTP payloads
We now ask Transmission to use gzip to compress the payload before sending it to
us. The savings are quite significant, as a torrent-get which is 2436 KiB in
plain text is compressed to just 215 KiB.

FPC includes a Pascal implementation of zlib so this comes with no extra
dependencies in the final binary.