Don't request labels if running with pre-3.0 Transmission
The labels field in torrent-get was added in Transmission 3.0, whose RPC version
is 16.

We request this field unconditionally, which breaks Reflection. I'm not sure how
pre-3.0 Transmission versions behave when they receive an unsupported field, but
we have a lot of these checks anyway already so it just makes sense do to it for
this one as well.